0.9.5
- Android example projects modernized to the latest Android Studio, SDK
and NDK developments. Including the experimental plugin for NDK.
- Removed seperate processing thread option from
SuperpoweredAndroidAudioIO, as it didn't help with any Android device
shipped with a bad HAL.
- SuperpoweredAdvancedAudioPlayer now has a displayPositionMs property,
which immediately reflects position changes after setPosition() or
seek() is called.
- The offline example project is extended with the
SuperpoweredOfflineAnalyzer.
- Fixed SuperpoweredTimeStretching bug in setting pitch shift cents
more than an octave up or down.

+++ Android/FrequencyDomain/app/src/main/java/com/superpowered/frequencydomain/MainActivity.java
@@ -8,7 +8,6 @@ import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
-
 public class MainActivity extends AppCompatActivity {
 
     @Override
@@ -53,5 +52,5 @@ public class MainActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
-    private native void FrequencyDomain(long samplerate, long buffersize);
+    private native void FrequencyDomain(int samplerate, int buffersize);
 }
